back/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
motor==3.4.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
deepface==0.0.92
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.3.0
tf-keras==2.16.0
pydantic==2.7.1
pydantic-settings==2.3.0
aiofiles==23.2.1